refact - fix oom issue - finall fix the issue
